chore: Remove circular import of PubNub dependencies
