+++ server/api/v0/join/index.ts
@@ -2,14 +2,14 @@ import type { User } from "../../../database/schemas";
 import { apiHandler, dispatchRequests } from "../../../helpers/apiHandler";
 import { BadRequestError } from "../../../errors/BadRequestError";
 import { DuplicateAccountError } from "../../../errors/DuplicateAccountError";
-import { generateHash, generateSalt } from "../../../auth/generators";
+import { generateAESCipherKey, generateHash, generateSalt } from "../../../auth/generators";
 import { is, nonempty, string, type } from "superstruct";
 import { MAX_USERS } from "../../../auth/limits";
 import { newAccessTokens, setSession } from "../../../auth/jwt";
-import { newPubNubCipherKey } from "../../../auth/pubnub";
 import { NotEnoughRoomError } from "../../../errors/NotEnoughRoomError";
-import { numberOfUsers, statsForUser, upsertUser, userWithAccountId } from "../../../database/io";
+import { numberOfUsers, statsForUser, userWithAccountId } from "../../../database/reads";
 import { respondSuccess } from "../../../responses";
+import { upsertUser } from "../../../database/writes";
 import { v4 as uuid } from "uuid";
 
 /**
@@ -49,7 +49,7 @@ export const POST = apiHandler("POST", async (req, res) => {
 	const passwordSalt = await generateSalt();
 	const passwordHash = await generateHash(givenPassword, passwordSalt);
 	const uid = newDocumentId();
-	const pubnubCipherKey = await newPubNubCipherKey();
+	const pubnubCipherKey = await generateAESCipherKey();
 	const user: Required<User> = {
 		currentAccountId: givenAccountId,
 		mfaRecoverySeed: null,
