+++ server/api/v0/leave/index.ts
@@ -1,11 +1,12 @@
 import { apiHandler, dispatchRequests } from "../../../helpers/apiHandler";
 import { BadRequestError } from "../../../errors/BadRequestError";
 import { compare } from "../../../auth/generators";
-import { destroyUser, userWithAccountId } from "../../../database/io";
+import { destroyUser } from "../../../database/writes";
 import { generateTOTPSecretURI, verifyTOTP } from "../../../auth/totp";
 import { is, nonempty, optional, string, type } from "superstruct";
 import { respondSuccess } from "../../../responses";
 import { UnauthorizedError } from "../../../errors/UnauthorizedError";
+import { userWithAccountId } from "../../../database/reads";
 
 export const POST = apiHandler("POST", async (req, res) => {
 	const reqBody = type({
