+++ server/auth/jwt.ts
@@ -1,8 +1,9 @@
 import type { JwtPayload, MFAOption, User } from "../database/schemas";
-import { addJwtToDatabase, jwtExistsInDatabase } from "../database/io";
+import { addJwtToDatabase } from "../database/writes";
 import { assertJwtPayload } from "../database/schemas";
 import { env, requireEnv } from "../environment";
 import { generateSecureToken } from "./generators";
+import { jwtExistsInDatabase } from "../database/reads";
 import { newPubNubTokenForUser, revokePubNubToken } from "./pubnub";
 import { ONE_HOUR } from "../constants/time";
 import Cookies from "cookies";
