+++ server/database/__mocks__/reads.ts
@@ -4,15 +4,9 @@ import type {
 	fetchDbCollection as _fetchDbCollection,
 	fetchDbDoc as _fetchDbDoc,
 	fetchDbDocs as _fetchDbDocs,
-	upsertUser as _upsertUser,
-	destroyUser as _destroyUser,
-	upsertDbDocs as _upsertDbDocs,
-	deleteDbDocs as _deleteDbDocs,
-	deleteDbDoc as _deleteDbDoc,
-	deleteDbCollection as _deleteDbCollection,
 	userWithUid as _userWithUid,
 	userWithAccountId as _userWithAccountId,
-} from "../io";
+} from "../reads";
 import { jest } from "@jest/globals";
 
 export const statsForUser = jest
@@ -34,17 +28,3 @@ export const fetchDbDoc = jest
 export const fetchDbDocs = jest
 	.fn<typeof _fetchDbDocs>()
 	.mockRejectedValue(new EvalError("This is a test"));
-
-export const upsertUser = jest.fn<typeof _upsertUser>().mockResolvedValue({ uid: "test_user_123" });
-
-export const destroyUser = jest.fn<typeof _destroyUser>().mockResolvedValue(undefined);
-
-export const upsertDbDocs = jest.fn<typeof _upsertDbDocs>().mockResolvedValue(undefined);
-
-export const deleteDbDocs = jest.fn<typeof _deleteDbDocs>().mockResolvedValue(undefined);
-
-export const deleteDbDoc = jest.fn<typeof _deleteDbDoc>().mockResolvedValue(undefined);
-
-export const deleteDbCollection = jest
-	.fn<typeof _deleteDbCollection>()
-	.mockResolvedValue(undefined);
